Learning about Sign In, Sign Up, Logout and so on.. using firebase authentication as well as best practices like async-await with redux-thunk, useRef to use react-navigation props in NavigationContainer.js

+++ screens/user/UserProductsScreen.js
@@ -1,5 +1,5 @@
-import React from 'react';
-import { FlatList, Button, Platform } from 'react-native';
+import React, { useState, useEffect } from 'react';
+import { Text, FlatList, Button, Platform, Alert, View, ActivityIndicator } from 'react-native';
 import { useSelector, useDispatch } from 'react-redux'
 import { HeaderButtons, Item } from 'react-navigation-header-buttons';
 
@@ -13,44 +13,75 @@ const UserProductsScreen = props => {
   const dispatch = useDispatch();
 
   const editProductHandler = (id) => {
-    props.navigation.navigate('EditProducts', {productId: id})
+    props.navigation.navigate('EditProducts', { productId: id })
   };
+
+  const [error, setError] = useState();
 
   const deleteHandler = (id) => {
     Alert.alert('Are you sure?', 'Do you really want to delete this item', [
-      {text: 'No', style: 'default'},
-      {text: 'Yes', style: 'destructive', onPress: () => {
-        dispatch(productsActions.deleteProduct(id));
-      }},
+      { text: 'No', style: 'default' },
+      {
+        text: 'Yes', style: 'destructive', onPress: async () => {
+          try {
+            await dispatch(productsActions.deleteProduct(id));
+          } catch (error) {
+            setError(error.message)
+          }
+        }
+      },
     ])
   };
 
-  return (<FlatList 
-    data={userProducts} 
-    keyExtractor={item => item.id} 
+
+  useEffect(() => {
+    if (error) {
+      Alert.alert('An error occurred!', error, [{ text: 'Okay', onPress: () => setError(null) }])
+    }
+  }, [error])
+
+  if (error) {
+    return (
+      <View>
+        <ActivityIndicator size="large" color={Colors.primary} />
+      </View>
+    );
+  }
+
+  if (userProducts.length === 0) {
+    return (
+      <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+        <Text>No products found, maybe start creating some?</Text>
+      </View>
+    )
+  }
+
+  return (<FlatList
+    data={userProducts}
+    keyExtractor={item => item.id}
     renderItem={itemData => (
-      <ProductItem 
-        image={itemData.item.imageUrl} 
-        title={itemData.item.title} 
-        price={itemData.item.price} 
+      <ProductItem
+        image={itemData.item.imageUrl}
+        title={itemData.item.title}
+        price={itemData.item.price}
         onSelect={() => {
           editProductHandler(itemData.item.id);
         }}
       >
-        <Button 
-          color={Colors.primary} 
-          title="Edit" 
+        <Button
+          color={Colors.primary}
+          title="Edit"
           onPress={() => {
             editProductHandler(itemData.item.id);
           }}
         />
-        <Button 
-          color={Colors.primary} 
-          title="Delete" 
+        <Button
+          color={Colors.primary}
+          title="Delete"
           onPress={() => deleteHandler(itemData.item.id)}
         />
       </ProductItem>
-    )}/>
+    )} />
   );
 };
 
@@ -58,28 +89,28 @@ UserProductsScreen.navigationOptions = navData => {
   return {
     headerTitle: 'Your Products',
     headerLeft: (
-    <HeaderButtons HeaderButtonComponent={HeaderButton} >
-      <Item 
-          title='Menu' 
-          iconName={Platform.OS === 'android' ? 'md-menu' : 'ios-menu'} 
+      <HeaderButtons HeaderButtonComponent={HeaderButton} >
+        <Item
+          title='Menu'
+          iconName={Platform.OS === 'android' ? 'md-menu' : 'ios-menu'}
           onPress={() => {
-              navData.navigation.toggleDrawer();
+            navData.navigation.toggleDrawer();
           }}
-      />
-    </HeaderButtons>
+        />
+      </HeaderButtons>
     ),
     headerRight: (
       <HeaderButtons HeaderButtonComponent={HeaderButton} >
-        <Item 
-            title='Add' 
-            iconName={Platform.OS === 'android' ? 'md-create' : 'ios-create'} 
-            onPress={() => {
-                navData.navigation.navigate('EditProducts');
-            }}
+        <Item
+          title='Add'
+          iconName={Platform.OS === 'android' ? 'md-create' : 'ios-create'}
+          onPress={() => {
+            navData.navigation.navigate('EditProducts');
+          }}
         />
       </HeaderButtons>
-      )
+    )
   }
-};  
+};
 
 export default UserProductsScreen;
